requirements.txt
flask==2.3.3
requests==2.31.0
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
